// src/classes/renderer/AnimatedTextureRenderer.d.ts
import * as PIXI from 'pixi.js';
import '../renderer/ContainerExtensions';
import RendererInterface from './RendererInterface';
import { ContainerExtensions } from '../renderer/ContainerExtensions';
import { ColorMatrixFilter } from '@pixi/filter-color-matrix';
export default class AnimatedTextureRenderer extends PIXI.Sprite implements RendererInterface, ContainerExtensions {
    private _imageNames;
    private _textures;
    private _currentFrame;
    lastAnim: number;
    animated: boolean;
    private _isPaused;
    loop: boolean;
    reversed: boolean;
    interval: number;
    pingPongMode: boolean;
    private _startFrame;
    private _endFrame;
    isOver: boolean;
    private _nextAnim;
    sheetName?: string;
    animationName?: string;
    constructor(imageNames?: string[], params?: {
        animated?: boolean;
        loop?: boolean;
        reversed?: boolean;
        interval?: number;
        pingPongMode?: boolean;
        sheetName?: string;
        animationName?: string;
        endFrame?: number;
        startFrame?: number;
        pause?: boolean;
        currentFrame?: number;
        randomFrame?: number;
        scale?: number | Point2D;
        scaleX?: number;
        scaleY?: number;
    } & Partial<Omit<PIXI.Sprite, 'scale'>> & Partial<RendererInterface>);
    hueFilter?: ColorMatrixFilter | undefined;
    blackAndWhiteFilter?: ColorMatrixFilter | undefined;
    saturationFilter?: ColorMatrixFilter | undefined;
    brightnessFilter?: ColorMatrixFilter | undefined;
    contrastFilter?: ColorMatrixFilter | undefined;
    grayscaleFilter?: ColorMatrixFilter | undefined;
    sleep: boolean;
    preventCenter?: boolean | undefined;
    _originalTexture?: PIXI.Texture<PIXI.Resource> | undefined;
    changeSheet(sheetName: string, animationName: string): void;
    get pause(): boolean;
    set pause(bool: boolean);
    get textures(): PIXI.Texture<PIXI.Resource>[];
    set textures(textures: PIXI.Texture<PIXI.Resource>[]);
    get currentFrame(): number;
    set currentFrame(val: number);
    get endFrame(): number;
    set endFrame(frame: number);
    get startFrame(): number;
    set startFrame(frame: number);
    get imageNames(): string[];
    set imageNames(names: string[]);
    update(): void;
    gotoAndPause(frame: number): void;
    stop(): void;
    restart(): void;
    onAnimEnd(): void;
    setTint(value: number): void;
    setHue(rotation: number, multiply: boolean): void;
    setBlackAndWhite(multiply: boolean): void;
    setSaturation(amount: number, multiply: boolean): void;
    setBrightness(b: number, multiply: boolean): void;
    setContrast(amount: number, multiply: boolean): void;
    setGreyscale(scale: number, multiply: boolean): void;
    setSize(width: number, height: number, preventCenter: boolean): void;
    setScale(x: number | {
        x: number;
        y: number;
    }, y?: number): void;
    center(): void;
    instantiate(params: any): void;
    static DEName: string;
}